services/web/requirements.txt
eventlet==0.30.2
Flask==2.1.1
Flask-Login
Flask-SocketIO==5.1.2
Flask-SQLAlchemy==2.5.1
#gevent
#gevent-websocket
simple-websocket
google-cloud-storage==2.4.0
gunicorn
icecream==2.1.2
ipdb==0.13.9
ipython==8.2.0
neovim==0.3.1
numpy==1.22.1
pandas==1.4.2
python-socketio==5.7.2
Pillow==9.1.0
pynvim==0.4.3
tqdm==4.64.0
wget==3.2
SQLAlchemy==1.4.36
